Fixed issues, and created heddoko library dll
Added the HeddokoLib library.
This library includes functions that are repeatedly used by many of the
heddoko apps, and as to maintain consistency across the board

Changed:
Removed Circular queue and using the HeddokoLib dll instead.  This dll
includes a fix to the circular buffer where atomic operations were not
being respected. The class should now be thread safe

Added in BodyView, TrackingThread, BodyFrameThread flags for pausing
thread operations
Once flagged, the operations are suspended until the flag is cleared.

Fixed an issue with BodySubsegmentView not reset to the initial frame