Updated clien
created views for in scene that connect to brianpack
create a brainpack controller whos responsibility is to interface with
unity to the brainpack
In TrackingThread:

Added a check in order to determine if the body being worked on has an
initial frame set. If not, set a trigger to the body view which then
safely sets the initialframe. Without this, unity didn't like that
another thread was updating one of the subsegments transforms.

In circular queue.cs
Enqueue: there was an issue enqueing data where the tail pointer was
going out of bounds. Fixed.

System.Reflection.TargetInvocationException
Fixed: this issue was being caused by a null pointer exception when
trying to pull data from the buffer.