Fixed the brainpack service connection
The  editor does add an extra overhead to the execution of the app so
some potential issues slip by and wouldn’t be noticed. The main cause
was because the server wasn’t being synchronized with the app, there
would be times when the socket was being closed by the server way before
the client had a chance to retrieve that data.
The fix is done by reworking the client to be synchronous, waiting for a
response from the server before closing the socket.